+++ docs/history.rst
@@ -1,10 +1,13 @@
 Version History
 ===============
 
-`1.2.0`_ Apr 1, 2019
+`2.0.0`_ Apr 1, 2019
 --------------------
-- Add ``history`` attribute of the response with all response objects
-- Add ``links`` attribute of the response with the parsed link header if set
+- Refactor the HTTPResponse to a stand-alone class
+  - Add ``history`` attribute of the response with all response objects
+  - Add ``links`` attribute of the response with the parsed link header if set
+  - Add ``exceptions`` attribute with stack of exceptions returned as responses
+- Add ``dont_retry`` as argument to ``http_fetch`` method
 - Change logging level in a few places to a more appropriate level
 - Add support for rejected consumers when auto-creating the ``User-Agent`` header
 - Add the netloc of a request to the log entry created when rate limited